revin-mini/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.2
openai==1.43.0

